Removed fromList & corrected bug in adjustComplete.
Function fromList always hung, so it was removed.

'adjustComplete' had a bug which made function 'update' ignore updates
to index 14 of the stream. To conceptualize the bug, imagine that we
have a 'Complete' made of a 'Bin' and two 'Tip's, and we call
'adjustComplete' with n=1. The result is that a recursive call is made
with n=-1 and the update is ignored.

Also some regular comments for some functions were turned into Haddock
comments.